Add PayPal token tradeoff on return from onboarding
Adds the ability to trade SharedId for an AccessToken, then
trade that for ClientId and ClientSecret, the final API credentials
that we'll use from this point onward.

This PR adds a dependency to PayPals Ruby Checkout SDK, however
the SDK does not support the token tradeoff that we need to do
initially, so I've added a Requests model that serves as a wrapper
for the SDK so I can add that functionality.

This PR also namespaces SolidusPaypalCommercePlatform for routing
purposes

Additionally, we're namespacing javascript functions, but because
the onboarding callback is handled by PayPals javascript, we're
unable to use the namespaced function there. We're instead
adding an ugly function name to the relevant view that maps to
the namespaced function.