Restore the generic bin/rails command
Move the sandbox and engine specializations to their own rails-*
files.

By default generate commands will be run in the context of the engine
while any other command will be run from the sandbox. With the
additional bin/rails-* files it will be more easy to find how to run
a command from a specific context.