Add frontend view and API controllers for PayPal checkout
Adds routes for orders_controller - which allows the creation of
orders on PayPals side. This is called when the paypal button is
first clicked, to send the order info to PayPal so they know how
to build the order and how much is going to be charged.
Also adds a route for payments_controller, which handles the creation
of the source/payment for the order.
en.yml was updated to display nothing for validation errors for
paypal_order_id - the reason is that when you try to proceed
without having created a payment on PayPal, the error message
"Payments Source #{en.yml entry} Can't Be Blank" appears, with
whatever you've set in the paypal_order_id entry appearing in the
middle of the string. We may want to adjust the entire error message,
but I think it's clearer with nothing in the middle - so just
"Payments Source Can't Be Blank", instead of "Payments Source
Paypal Order Id Can't Be Blank".
Also adds tests for the frontend! :tada: