Add ability for admin users to customize paypal button
I utilize the preference_fields, creating a custom one for a select option. I also provide defaults for
all of the button options based on the defaults PayPal recommends selecting, while adding a helper
that provides the collection of other available options. All of this is displayed on the PCP
payment method page on admin. The options are passed on the frontend when we display the PayPal button.