Add Order Simulator to simulate address changes
When the user changes their address on PayPal, PayPal only tells us
part of the new address. We don't want to update the address on file,
since it's only a partial and could lead to a mismatched address if
the user decides not to pay with PayPal.

To solve this, I've created the Order Simulator - it dups the order
and adds the partial shipping address, allow us to see what the
shipping rates and taxes _would_ be for that partial address. None of
the simulated objects are saved.