Allow paypal_address to receive recipients name & email
PayPal also passes us the recipients name and email. We want to use the name for the
shipping address, and use the email if the order doesn't already have an email
associated with it. Also allows paypal_address to create a new address if the order
doesn't already have one - which is useful for allowing checkout on cart and product pages,
where the user won't necessarily have an address on their order.