Turn off phone number required validation for Spree::Address
Since PayPal doesn't collect this information, and we're using PayPal for checkout
in cases where the customer checks out on the cart or product page, I turned off
phone number required validation. I also added a bit to the README that details
how to turn it back on, if the user needs this functionality.

This seemed like the best solution to the problem - The other solutions involved
collecting phone numbers from the user before or after the PayPal checkout process,
but it seemed to be a bit too unusual for the user to come back from paying on PayPal
only to be asked for their phone number too.

With issue solidusio/solidus#3679 it looks like Solidus phone number validations are
going to default to not required anyway, so hopefully this solution isn't too out of the
norm!