Add PayPal button to the cart page
Adds a deface file that loads in cart buttons for the cart page. The cart_buttons partial
is set up to handle multiple cart buttons. Within the context of this gem, the user can have
multiple PayPal Commerce Platform payment methods that display on the cart (though I don't
know why they would want to) - but in the future, other paymetn methods could also add
buttons to the cart page. We may want to move this code to Solidus if that's the case.