Add API endpoint for creating a new order
Adds an API endpoint to control order creation from the product PayPal button.
The users existing order - if any - is replaced with this new order. This order
should contain only the chosen quantity of the selected variant on that product
page. It also associates the current user with the order, if it can find them.