Verify PayPals total and ensure order is fully updated
Adds a function that hits the amount verification API endpoint,
so that we can be sure that the amount PayPal sends us is the
correct amount to charge the user.

Also moves the advanceOrder function to before the addPayment
function, since we need the shipment_total to be updated and
that only happens on the shipment order state.