Send amount of payment to PayPal on capture
Previously, we were just telling PayPal to capture the amount that
was authorized. However, admin users have the ability to edit the
amount of the payment, so now we're sending the amount of the payment
to PayPal for capturing.

You can capture less than the authorized amount just fine, however
trying to capture MORE than that amount will fail the payment and
flash a "Something went wrong" error. I think that's fine for the
functionality we want - I would expect to not be able to charge
more than the customer authorized.

Should fix #29