Include payment method information in pricing options cache key
The cache for products needs to be expired when the paypal commerce platform
payment method is changed, deleted, or updated - since it displays a button
on the product page. This adds a custom pricing_options_class - which just
updates the cache_key to include SPCP payment method information. Thanks
to @elia for actually doing this work - I just adapted it to this extension.
:pray: