Use options instead of preferences
If the user is using a static preference model, `payment_method.preferences`
returns the static preference model instead of a hash of preferences.
`payment_method.options` always returns a hash of preferences, so we want to
use that instead.

Also adds some defaults for some of the preference getters, because if they
aren't specifically set in the static preference model, they will return
nil instead of the default value.