Change outstanding_balance to total in order total checker
Previously we were checking PayPals total against the orders
outstanding balance, however if the outstanding balance is
different than the order total, this check would never pass
because we send PayPal the whole order. Changing this to total
instead is a bit more clear and provides less room for errors.