Replace order_simulator with ActiveRecord transaction
Previously when we got a partial address from PayPal, we would dup
the order, and use it in its unsaved state to simulate the new
address, so nothing would be actually changed or added to the DB.
However, this does not play well with a lot of Solidus methods,
like promotion activation (because it calls `create!` which doesn't
work with unsaved objects). It was also becoming a bit too complex.

This PR moves to simplify all of that by just wrapping the address
change in an ActiveRecord transaction and rolling it back after we
get the data we need to send back to PayPal.

In essence, this will change the orders ship_address, generate new
shipments, and format all of that to be sent to PayPal, and then undo
all of the actions it just took to generate that data, leaving the order
untouched at the end.