Add better error handling to button_actions.js
Previously, if an error occurred in the send_order function - like if
the order does not total up correctly - the user would see the PayPal
loading screen forever. This handles the error - closing the PayPal screen
when they occur - and logging the error + debug ID both to console and to
the Rails logger. This also ensures that an address is rejected if there is
an issue sending the updated order details to PayPal, otherwise PayPal would
allow paying for an order with address issues.