Add a state_guesser class to guess states
Solidus in some cases uses the incorrect states for a country - for instance,
Solidus uses the first level of subregions in Carmen for Italy, while it
actually needs to use the second level of subregions. This causes issues
with this extension, as if the PayPal address a user has is in Italy - or
one of the other states with this issue - the address validation will fail,
and PayPal will tell the user that we don't ship to that area, essentially
preventing checkout in these cases.

This adds a state guesser class that attemps to help in those situations by
finding the Carmen region, and returning the regions parent region, which
should match the Solidus state on file. I'm not sure if this will work in
every country with this issue, but it at least solves the issue with Italy.