Remove partner variable information from README
You wouldn't normally change these, and they only matter in the
context of the wizard - they're not used elsewhere. Unless you're
building a marketplace, I don't think there's any reason to adjust
these. In addition, we already list PAYPAL_CLIENT information which
conflicts with this a bit, so I think to clear up the confusion
we should just remove this section from the README. Ref: #97