Allow control over Venmo PayPal financing option
Resolves issue/ticket:
- https://github.com/solidusio-contrib/solidus_paypal_commerce_platform/issues/136

If a transaction meets Venmo prerequisites [1], then Venmo button
should appear already on checkout, cart and product page depending
on payment method preferences.

These changes is to give more control to the developer so that they can
disable Venmo by setting the `enable_venmo` to false.
https://developer.paypal.com/docs/business/javascript-sdk/javascript-sdk-configuration/#disable-funding

As it Venmo may have already been showing for developers, the default
was set to true.

The enable-funding option is now also sent over to ensure that the Venmo
button gets rendered if eligible:
https://developer.paypal.com/docs/business/javascript-sdk/javascript-sdk-configuration/#enable-funding
https://developer.paypal.com/docs/business/checkout/pay-with-venmo/integrate-pay-with-venmo/#always-request-venmo

[1] https://developer.paypal.com/docs/business/checkout/pay-with-venmo/#eligibility