Add force_buyer_country payment method preference for testing
Resolves issue/ticket:
- https://github.com/solidusio-contrib/solidus_paypal_commerce_platform/issues/137

Usually on checkout, PayPal will look at the buyer's location by their
ip and determine which funding sources are available to them. For
example, allow Venmo as an option for US buyers, but not for others.

With this new preference, you can set a country to override the buyer
location. Allowing you on non-production environments to see the
different funding sources for different customers and its integration.

This preference is still available on production but has no effect.
It was kept on production to not have differences between the
environments. Otherwise, errors could easily occur on production if
for example, preferred_force_buyer_country method was incorrectly
called.

More information:
https://developer.paypal.com/docs/business/javascript-sdk/javascript-sdk-configuration/#buyer-country