Update Venmo integration
The Venmo integration now adds a standalone button for Venmo, rather
than enabling Venmo as a financing option for the normal checkout flow.
This is because Venmo does not currently support shipping callbacks, so
we cannot use this.

The Venmo integration adds a Venmo button to payment page. It cannot
be rendered on the cart or product page, because it is not possible in
Venmo to get shipping costs from Solidus and update for the payment
total.