Add fundingType to PaypalCommercePlatformSources
Saving the data which fundingSource, i.e. which financing option was
used for the PayPal payment. This data be shown on the frontend to
confirm to the user how they paid, for example, Venmo.

This is reassurance and was also a requirement set by PayPal for our
integration of Venmo.