Block Capture on PayPal PENDING Payments
In bd079eff1049c0865913d948d2dc4ba460d7dbf6 we tried to block the capture
of payments by checking the status in #can_capture? However at this stage
there isn't yet a payment created in PayPal so the added condition will
always return true.

Instead with this commit we are moving the check logic to the #capture
method itself, returning a failed ActiveMerchant::Billing::Response if
the PayPal payment is pending.