Use current code in sandbox generation
With this change we can use the current code of the extension
instead of the one picked by the starter frontend installer by
default.

This will streamline development, being able to reflect changes
to the sandbox immediately (regenerating it without other tweaks).

Also, this has the side effect to keep working if we change the
payment method default in the future.