Allow to use the auto_capture global preference
In the Payment Method edit UI, we can choose the autocapture
preference we want to be applied for their payments.

The choices are: "Yes", "No", "Use App Default".

When the last option is picked, the value for this boolean is
set to nil and we were not considering this last option through
the codebase.

Luckily, a convenient method in Solidus [1] allows us to return
the App Default value.

[1]: https://github.com/solidusio/solidus/blob/929035c957f75ebd7ae745af69bf2a93a3727a87/core/app/models/spree/payment_method.rb#L156-L158